fake news/requirements.txt
streamlit==1.28.1
python-dotenv==1.0.0
openai==1.55.0
